pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.54.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "openai>=1.86.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "stripe>=12.2.0",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
]
